Assignment # 3 Ervyn Cruz
I used a while-loop statement which produce an infinite loop of 3 digit numbers where it produced the statement of "Is Ascending" "Is Descending" or "Is neither descending nor ascending". I also produced a count statement which counts the integer up until 30 where it would stop the processing of the integers as it got to 30. I changed a few things where before they had x in terms of 'number' and changed it to as the variable x and also changed the statement where instead of the integer saying "Ascending" etc., it would say " The number is ascending, descending, neither."